Update jenkinsfile
parameters: [text(name:'plan',description:'please review the plan',defaultValue:plan)]

parameters:: This is a Jenkins Pipeline directive used to define input parameters for a build. When you run a Jenkins job that has parameters, Jenkins will display a form where users can input values before starting the build.

[ ... ]: This indicates a list of parameters. You can define multiple parameters within this block.

text(...): This specifies the type of parameter. In this case, text creates a multi-line text input field in the Jenkins UI. This is useful for accepting longer strings of text, like a detailed plan or configuration.

Inside the text() function, there are three arguments:

name:'plan': This sets the name of the parameter. This name (plan) is how you'll refer to and access the value of this parameter later in your Jenkins Pipeline script (e.g., params.plan).

description:'please review the plan': This provides a description that will be displayed next to the text input field in the Jenkins UI. It tells the user what this parameter is for. In this case, it's instructing the user to review the plan.

defaultValue:plan: This sets the default value for the parameter. When the Jenkins job is run, this field will be pre-filled with the content of a variable named plan.

In the context of a Terraform pipeline, this plan variable would typically hold the output of a terraform plan command, which shows the proposed infrastructure changes. This allows the user to see the full Terraform plan directly in the Jenkins UI before deciding whether to proceed with terraform apply.
In essence, this command does the following:

It configures a Jenkins job to have a parameter named "plan" which is a multi-line text box. This text box will be pre-populated with the content of the plan variable (likely a Terraform execution plan) and will have the helpful label "please review the plan" for the user. This is a common pattern in CI/CD pipelines for Terraform, allowing for manual review and approval of infrastructure changes.